requirements.txt
altgraph==0.17
appdirs==1.4.4
astroid==2.6.5
auto-py-to-exe==2.9.0
black==21.7b0
bottle==0.12.19
bottle-websocket==0.2.9
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.6
click==7.1.2
colorama==0.4.4
cssselect2==0.4.1
cycler==0.10.0
defusedxml==0.7.1
Eel==0.12.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
html5lib==1.1
isort==5.9.2
Jinja2==3.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.1
pandas==1.3.1
pathspec==0.9.0
pefile==2021.5.24
Pillow==8.3.1
pipdeptree==2.0.0
pycparser==2.20
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pylint==2.9.5
pyparsing==2.4.7
pyphen==0.11.0
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywin32-ctypes==0.2.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
regex==2021.7.6
six==1.16.0
tinycss2==1.1.0
toml==0.10.2
tomli==1.1.0
WeasyPrint==52.5
webencodings==0.5.1
whichcraft==0.6.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
